Add logging and dabbing around in OpenCL

+++ BlurredLines.Tests/Calculation/GaussianBlurKernelCalculatorTests.cs
@@ -18,7 +18,7 @@ namespace BlurredLines.Tests.Calculation
 
         [Theory]
         [MemberData(nameof(ExpectedOneDimensionalKernelValues))]
-        public void ItShallCalculateCorrectOneDimensionalKernelValues(int kernelSize, double[] expectedKernel)
+        public void ItShallCalculateCorrectOneDimensionalKernelValues(int kernelSize, float[] expectedKernel)
         {
             // When
             var actualKernelValues = sut.CalculateOneDimensionalKernel(kernelSize);
@@ -27,7 +27,7 @@ namespace BlurredLines.Tests.Calculation
             actualKernelValues.Length.Should().Be(expectedKernel.Length);
             for (int i = 0; i < actualKernelValues.Length; i++)
             {
-                actualKernelValues[i].Should().BeApproximately(expectedKernel[i], 0.0001);
+                actualKernelValues[i].Should().BeApproximately(expectedKernel[i], 0.0001f);
             }
         }
 
@@ -48,12 +48,12 @@ namespace BlurredLines.Tests.Calculation
         public static IEnumerable<object[]> ExpectedOneDimensionalKernelValues =>
             new List<object[]>
             {
-                new object[] {3, new[] {0.27901, 0.44198, 0.27901}},
-                new object[] {5, new[] {0.06136, 0.24477, 0.38774, 0.24477, 0.06136}},
-                new object[] {7, new[] {0.00598, 0.060626, 0.241843, 0.383103, 0.241843, 0.060626, 0.00598}},
+                new object[] {3, new[] {0.27901f, 0.44198f, 0.27901f}},
+                new object[] {5, new[] {0.06136f, 0.24477f, 0.38774f, 0.24477f, 0.06136f}},
+                new object[] {7, new[] {0.00598f, 0.060626f, 0.241843f, 0.383103f, 0.241843f, 0.060626f, 0.00598f}},
                 new object[]
                 {
-                    9, new[] {0.000229, 0.005977, 0.060598, 0.241732, 0.382928, 0.241732, 0.060598, 0.005977, 0.000229}
+                    9, new[] {0.000229f, 0.005977f, 0.060598f, 0.241732f, 0.382928f, 0.241732f, 0.060598f, 0.005977f, 0.000229f}
                 },
             };
     }
